% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_functions.R
\name{censor_photos}
\alias{censor_photos}
\title{Remove UWIN Images With Faulty Timestamps}
\usage{
censor_photos(uwin_data = NULL)
}
\arguments{
\item{uwin_data}{The list object returned from \code{\link{collect_tables}}.}
}
\value{
Returns the list object, but with offending photos and detections
  removed from their respective tables.
}
\description{
\code{censor_photos} removes photos from the \code{'Photos'}
  table and associated detections from the \code{'Detections'} table if the
  time stamp on the photo is >30 days before a camera was set at a
  site or > 30 days after a camera was pulled at a site. These latter values
  are stored in the \code{'Visits'} table.
}
\details{
This function requires the following tables in the list object to
  work : \code{'Photos'}, \code{'Detections'}, and \code{'Visits'}.

  Furthermore, \code{\link{do_qaqc}} must be applied to the uwin list object
  before photos can be summarised.
}
\examples{
 # read in the data
 uwin_list <- collect_tables("UWIN_DB_CHIL.accdb")

 # check for errors
 uwin_list <- do_qaqc(uwin_list)

 # censor the photos
 uwin_list <- censor_photos(uwin_data = uwin_list)

}
