% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_fun.R
\name{collect_tables}
\alias{collect_tables}
\title{Collect tables from UWIN database}
\usage{
collect_tables(database = NULL, tables = NULL)
}
\arguments{
\item{database}{File name of the UWIN Access database as a character vector.
If \code{database} does contain an absolute path, the file name is relative
to the current working directory.}

\item{tables}{A character vector of the table names to be pulled from the
UWIN database. If this argument is left blank than \code{tables} is set to
\code{c("CameraLocations", "Detections", "Photos", "Species",
"StudyAreas", "Visits", "lkupAction", "lkupSeasons")}.}
}
\value{
A named list of tables from the UWIN database. Each table will be
  returned as a \code{\link{data.table}} instead of a
  \code{\link{data.frame}}, as this considerably speeds up
  summarizing these data.
}
\description{
\code{collect_tables} uses the \code{\link{RODBC}} package to connect
  to the UWIN database and collect the requested tables.

  Note: For this function to work the Microsoft Access Database Engine
  must be installed. It can be found
  \href{https://www.microsoft.com/en-us/download/details.aspx?id=13255}{here}.
}
\examples{
dat <- collect_tables("UWIN_DB_CHIL.accdb")
dat <- collect_tables("UWIN_DB_CHIL.accdb", tables = c("Photos", "Visits"))

}
\author{
Mason Fidino
}
