% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_functions.R
\name{create_detection_matrix}
\alias{create_detection_matrix}
\title{Create species detection non-detection matrix}
\usage{
create_detection_matrix(uwin_data = NULL, observation_matrix = NULL,
  binomial_detections = FALSE, select_species = TRUE, species = NULL)
}
\arguments{
\item{uwin_data}{The list object returned from \code{\link{collect_tables}}
and after it has been through \code{\link{do_qaqc}} and reduced to the
seasons of interest via \code{\link{reduce_seasons}}.}

\item{observation_matrix}{The list object returned by
\code{\link{create_observation_matrix}}.}

\item{binomial_detections}{If \code{TRUE}, \code{create_detection_matrix}
will return the total number of days a species was observed at a site.
If \code{FALSE}, \code{create_detection_matrix} will return a vector of
binary elements that take the value of \code{1} if a species was observed
on a given day, \code{0} if it was not, or \code{NA} if the camera was
not operable.}

\item{select_species}{If TRUE, a pop-up list will open up that you can use
to select the species you would like to make a detection matrix for. You
can hold \code{Ctrl} to select multiple species that are seperated by
other species you do not want to make a detection matrix for. Defaults
to TRUE.}

\item{species}{A vector of the species names from the \code{ShortName}
column of the \code{Species} table within the UWIN database. If left NULL
and \code{select_species = FALSE} then a detection matrix
will be made for each species in the \code{Species} table.}
}
\value{
A list with three elements. The first element, \code{mat}, contains
  a survey ID (i.e., site-season-year abbreviation). The second element,
   \code{days_active}, is a vector of the days that camera traps were active
   in a given season. The final element, \code{binom_mat} is the total number
   of days a camera trap was active on a given season.
}
\description{
Create species detection non-detection matrix
}
\examples{

 # read in the data
 dat <- collect_tables("UWIN_DB_CHIL.accdb")

 # apply qaqc
 dat <- do_qaqc(dat)

 # collect only one season of data
 dat <- reduce_seasons(dat, start = "JU17")

 # make observation matrix
 obser_matrix <- create_observation_matrix(dat)

 # make a detection matrix
 detect_matrix <- create_detection_matrix(dat, obser_matrix)

}
