% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_functions.R
\name{create_observation_matrix}
\alias{create_observation_matrix}
\title{The number of days each camera trap is operable per season}
\usage{
create_observation_matrix(uwin_data = NULL, drop_tails = FALSE)
}
\arguments{
\item{uwin_data}{The list object returned from \code{\link{collect_tables}}
and after it has been through \code{\link{do_qaqc}} .
If the \code{Visits} table is not within this object an error will occur.}

\item{drop_tails}{This will check if the date range for a site taken from
the camera trap images occurs between when the camera set date and camera
pull date entered into the \code{Visits} table of the UWIN database. If
the date range is > 7 days before the first recorded camera set for a
sampling season or > 7 days after the last recorded camera pull for a
sampling season then those days will be removed from the analysis.}
}
\value{
A list with 3 elements. The first element, \code{mat}, contains
  either a survey ID (i.e., site-season-year abbrevaition) by date matrix
  if \code{binomial_detections = FALSE} or a named vector of the number of
  days each survey ID was active. The second element, \code{days_active},
  is a vector of the days that camera traps were active in a given season.
  The final element, \code{binom_mat}, is a binomial version of the
  \code{mat} element (i.e., the rowSums from \code{mat}).
}
\description{
The number of days each camera trap is operable per season
}
\examples{

 # read in the data
 # not run: dat <- collect_tables("UWIN_DB_CHIL.accdb")

 # apply qaqc
 dat <- do_qaqc(dat)

 # collect only one season of data
 dat <- reduce_seasons(dat, start = "JU17")

 # make observation matrix
 obser_matrix <- create_observation_matrix(dat)
}
