% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc_functions.R
\name{photos_qaqc}
\alias{photos_qaqc}
\title{Check 'Photos' table for errors}
\usage{
photos_qaqc(uwin_data = NULL, file_conn = NULL)
}
\arguments{
\item{uwin_data}{The list object returned from \code{\link{collect_tables}}.
If the \code{Photos} and \code{Visits} table is not within this object
an error will occur.}

\item{file_conn}{The file path in which to write errors to supplied
  as a character string. This argument is managed automatically if
  \code{\link{do_qaqc}} is called instead. If left \code{NULL} then
  \code{visits_qaqc} will create a \code{error_reports} sub-folder
  in the working directory and populate it with an error report titled
\code{error_report_DATE.txt} where
\code{DATE} is the current date called via \code{\link{Sys.Date}}}
}
\description{
\code{photos_qaqc} looks for errors in the timestamps of
  the photos uploaded into the UWIN database. The 'Visits' table must
  also be present in the list object supplied by \code{\link{collect_tables}}
  as the timestamps are compared to the set and pull records in this table.
}
\examples{
# read in the data
uwin_list <- collect_tables("UWIN_DB_CHIL.accdb")

# conduct qaqc on 'Photos' table
uwin_list <- photos_qaqc(uwin_list)
}
\author{
Mason Fidino
}
