% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{reduce_seasons}
\alias{reduce_seasons}
\title{Query data within two sampling periods}
\usage{
reduce_seasons(uwin_data = NULL, start = NULL, end = NULL)
}
\arguments{
\item{uwin_data}{The list object returned from \code{\link{collect_tables}}.}

\item{start}{A character vector that contains the first sampling period
and year that you would like to query.
\code{start} must be 4 characters long with the first two characters denoting
the sampling season and the last two characters denoting the year
(e.g., April 2017 would be \code{"AP17"}).}

\item{end}{A character vector that contains the last sampling period and year
that you would like to query.
\code{start} must be 4 characters long with the first two characters denoting
the sampling season and the last two characters denoting the year
(e.g., April 2017 would be \code{"AP17"}).
If only querying one season of data \code{end} should be left as \code{NULL}.}
}
\value{
Returns the list object from \code{\link{collect_tables}} with data
  from either a single season (if \code{end} is left NULL) or with data
  that lie between the sampling periods specified in \code{start} and
  \code{end}. Note that this only queries data in the Visits, Photos, and
  Detections table within the Access database.
}
\description{
\code{reduce_seasons} queries data for a single season or between
two seasons depending on how arguments are filled.
}
\examples{
dat <- collect_tables("UWIN_DB_CHIL.accdb")

dat <- reduce_seasons(dat, start = "JA16", end = "JU17")

# if only collecting data from one season.

dat <- collect_tables("UWIN_DB_CHIL.accdb")

dat <- reduce_seasons(dat, start = "JA16")

}
\author{
Mason Fidino
}
