% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc_functions.R
\name{visits_qaqc}
\alias{visits_qaqc}
\title{Check visits table for errors}
\usage{
visits_qaqc(uwin_data = NULL, stop_on_error = TRUE)
}
\arguments{
\item{uwin_data}{The list object returned from \code{\link{collect_tables}}.
If the \code{Visits} table is not within this object an error will occur.}

\item{stop_on_error}{A logical statement to determine if this function
stops when it finds any errors in the Visits table. Currently defaults to
\code{TRUE}.}
}
\value{
Returns \code{uwin_data}, will stop if data is not correctly entered
  and \code{stop_on_error} is \code{TRUE}.

  \code{visits_qaqc} also creates a new column in the visits table titled
  \code{SurveyID} which concatenates the \code{LocationID}, \code{SeasonID},
  and year from \code{VisitDate}, seperated by a dash (e.g.,
  \code{"449-4-17"}).
}
\description{
\code{visits_qaqc} looks for data entry errors in the visits table
 in the UWIN database. This currently checks if the same action occurs
 twice within a single visit, or if a camera is 'set' multiple
 times in a single season.
}
\examples{
uwin_list <- visits_qaqc(uwin_data = uwin_list)

# This function even works without assigning the object
visits_qaqc(uwin_data = uwin_list)
}
\author{
Mason Fidino
}
