% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc_functions.R
\name{visits_qaqc}
\alias{visits_qaqc}
\title{Check 'Visits' table for errors}
\usage{
visits_qaqc(uwin_data = NULL, file_conn = NULL)
}
\arguments{
\item{uwin_data}{The list object returned from \code{\link{collect_tables}}.
If the \code{Visits} table is not within this object an error will occur.}

\item{file_conn}{The file path in which to write errors to supplied
  as a character string. This argument is managed automatically if
  \code{\link{do_qaqc}} is called instead. If left \code{NULL} then
  \code{visits_qaqc} will create a \code{error_reports} sub-folder
  in the working directory and populate it with an error report titled
\code{error_report_DATE.txt} where
\code{DATE} is the current date called via \code{\link{Sys.Date}}}
}
\value{
Returns \code{uwin_data}, will stop if data is not correctly entered
  and \code{stop_on_error} is \code{TRUE}.

  \code{visits_qaqc} also creates a new column in the visits table titled
  \code{SurveyID} which concatenates the \code{LocationID}, \code{SeasonID},
  and year from \code{VisitDate}, seperated by a dash (e.g.,
  \code{"449-4-17"}).
}
\description{
\code{visits_qaqc} looks for data entry errors in the 'Visits' table
  in the UWIN database.
}
\examples{
# Load data from database
# not run: uwin_list <- collect_tables("UWIN_DB_CHIL.accdb")

# apply qaqc to 'Visits' table
uwin_list <- visits_qaqc(uwin_data = uwin_list)

}
\author{
Mason Fidino
}
